backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework-gis==1.0
# gdal is installed via system package (python3-gdal) in Dockerfile
# Do not install via pip as it requires compilation tools
